+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to
 
 ### Added
 
+- Retry webhook events on transient database connection errors
+  [#3097](https://github.com/OpenFn/lightning/issues/3097)
+
 ### Changed
 
 - Enable X-Content-Type-Options header for static pages.
@@ -40,8 +43,8 @@ and this project adheres to
 - Add test gauge metric that can be used to set arbitrary values for the
   purposes of triggering behaviour in metric consumers.
   [3510](https://github.com/OpenFn/lightning/issues/3510)
-- Add test gauge metric that can be used to set arbitrary values for the purposes
-  of triggering behaviour in metric consumers.
+- Add test gauge metric that can be used to set arbitrary values for the
+  purposes of triggering behaviour in metric consumers.
   [#3510](https://github.com/OpenFn/lightning/issues/3510)
 - Possibly temporary plumbing to allow the use of libcluster_postgres as an
   additional mechanism for discovering Erlang nodes.